# docs/api/openapi.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Concordat.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/schemas/{id}": {
      "get": {
        "tags": [
          "Schemas"
        ],
        "summary": "Get a schema by its content-addressed id",
        "description": "Schemas are stored globally and deduplicated by content (ADR-015), so there is no tenant column to filter on. Access is authorised by reachability: a caller may read a schema only if some subject in their tenant references it. An unreachable id returns 404, identical to one that does not exist — a distinct 403 would confirm another tenant's schema exists.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/schemas/{id}/subjects": {
      "get": {
        "tags": [
          "Schemas"
        ],
        "summary": "List the subjects and versions using a schema",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SchemaUsage"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/schemas/{id}/bundled": {
      "get": {
        "tags": [
          "Schemas"
        ],
        "summary": "Get a schema as one self-contained document",
        "description": "Inlines every Concordat reference, transitively, into $defs and rewrites the refs as local pointers, so the result validates without further fetches. Bundling happens here rather than at registration: doing it at registration would make canonicalisation depend on registry state, and no SDK could then reproduce a schema id offline.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundledSchemaResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/schemas/lookup": {
      "post": {
        "tags": [
          "Schemas"
        ],
        "summary": "Find the id a schema document would receive",
        "description": "Canonicalises the document and returns its content-addressed id, plus whether it is already known. Never writes.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupSchemaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupSchemaResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BundledSchemaResponse": {
        "required": [
          "schemaId",
          "format",
          "bundled",
          "inlined"
        ],
        "type": "object",
        "properties": {
          "schemaId": {
            "type": "string",
            "description": "The id of the root schema."
          },
          "format": {
            "type": "string",
            "description": "The schema language."
          },
          "bundled": {
            "type": "string",
            "description": "The document, with every reference inlined into `$defs`."
          },
          "inlined": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The references that were inlined."
          }
        },
        "description": "A self-contained schema."
      },
...
      "LookupSchemaRequest": {
        "required": [
          "format",
          "schema"
        ],
        "type": "object",
        "properties": {
          "format": {
            "type": "string",
            "description": "One of `json`, `avro`, `protobuf`."
          },
          "schema": {
            "type": "string",
            "description": "The schema document."
          }
        },
        "description": "Request to look a schema up by its content."
      },
      "LookupSchemaResponse": {
        "required": [
          "schemaId",
          "known",
          "canonical"
        ],
        "type": "object",
        "properties": {
          "schemaId": {
            "type": "string",
            "description": "The id this document would receive."
          },
          "known": {
            "type": "boolean",
            "description": "Whether the caller can already reach a stored schema with this id."
          },
          "canonical": {
            "type": "string",
            "description": "The canonical text the id was computed from."
          }
        },
        "description": "The answer to a schema lookup."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ReferenceResponse": {
        "required": [
          "name",
          "subject",
          "version"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The `$ref` text as it appears in the document."
          },
          "subject": {
            "type": "string",
            "description": "The referenced subject."
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The referenced version ordinal.",
            "format": "int32"
          }
        },
        "description": "One outgoing reference."
      },
...
      "SchemaResponse": {
        "required": [
          "schemaId",
          "format",
          "schema",
          "references"
        ],
        "type": "object",
        "properties": {
          "schemaId": {
            "type": "string",
            "description": "The content-addressed id."
          },
          "format": {
            "type": "string",
            "description": "The schema language."
          },
          "schema": {
            "type": "string",
            "description": "The canonical text."
          },
          "references": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReferenceResponse"
            },
            "description": "Outgoing references, derived from the document."
          }
        },
        "description": "A stored schema."
      },
      "SchemaUsage": {
        "required": [
          "subject",
          "version"
        ],
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "description": "The subject name."
          },
          "version": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The version ordinal.",
            "format": "int32"
          }
        },
        "description": "One place a schema is in use."
      },
...
    }
  },
...
}